% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\docType{class}
\name{Cluster}
\alias{Cluster}
\title{Cluster Object}
\format{
An \code{\link{R6Class}} generator object
}
\description{
An R6 class defining a cluster object for parallel computations
}
\author{
Daniel Hohmann
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cluster$new()}}
\item \href{#method-export}{\code{Cluster$export()}}
\item \href{#method-loadPackages}{\code{Cluster$loadPackages()}}
\item \href{#method-apply}{\code{Cluster$apply()}}
\item \href{#method-stop}{\code{Cluster$stop()}}
\item \href{#method-clone}{\code{Cluster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new cluster object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$new(numberOfCores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{integer; the number of cores to be used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export objects to the cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$export(objects)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objects}}{list; objects to be exported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadPackages"></a>}}
\if{latex}{\out{\hypertarget{method-loadPackages}{}}}
\subsection{Method \code{loadPackages()}}{
Load packages for the cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$loadPackages(packages)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{packages}}{character vector; the packages to be loaded}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
Load balanced parallel execution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$apply(args, fun, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{vector of list of arguments to be passed to fun}

\item{\code{fun}}{function; the function to be executed}

\item{\code{...}}{additional arguments to be passed to fun}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the cluster connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cluster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
