% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIOPA_RFRs.R
\docType{class}
\name{EIOPA_RFR}
\alias{EIOPA_RFR}
\title{EIOPA RFR}
\format{
An \code{\link{R6Class}} generator object
}
\description{
Read and use EIOPA risk free rates
}
\author{
Daniel Hohmann
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EIOPA_RFR$new()}}
\item \href{#method-available_RFRs}{\code{EIOPA_RFR$available_RFRs()}}
\item \href{#method-get_RFR_table}{\code{EIOPA_RFR$get_RFR_table()}}
\item \href{#method-get_RFR_by_country_code}{\code{EIOPA_RFR$get_RFR_by_country_code()}}
\item \href{#method-forward_rates}{\code{EIOPA_RFR$forward_rates()}}
\item \href{#method-availability_overview}{\code{EIOPA_RFR$availability_overview()}}
\item \href{#method-clone}{\code{EIOPA_RFR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new EIOPA RFR object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to the input excel file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-available_RFRs"></a>}}
\if{latex}{\out{\hypertarget{method-available_RFRs}{}}}
\subsection{Method \code{available_RFRs()}}{
Get list of available EIOPA RFRs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$available_RFRs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\linkS4class{tbl_df} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_RFR_table"></a>}}
\if{latex}{\out{\hypertarget{method-get_RFR_table}{}}}
\subsection{Method \code{get_RFR_table()}}{
Export the complete RFR table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$get_RFR_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\linkS4class{tbl_df} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_RFR_by_country_code"></a>}}
\if{latex}{\out{\hypertarget{method-get_RFR_by_country_code}{}}}
\subsection{Method \code{get_RFR_by_country_code()}}{
Export a specific RFR by country code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$get_RFR_by_country_code(country)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country}}{a string specifying the Country code}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\linkS4class{tbl_df} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forward_rates"></a>}}
\if{latex}{\out{\hypertarget{method-forward_rates}{}}}
\subsection{Method \code{forward_rates()}}{
Calculate monthly forward rates from EIOPA ZSKs using Smith-Wilson-Interpolation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$forward_rates(country, months = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country}}{a string specifying the Country code}

\item{\code{months}}{an integer vector specifying the relevant forward terms in months (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-availability_overview"></a>}}
\if{latex}{\out{\hypertarget{method-availability_overview}{}}}
\subsection{Method \code{availability_overview()}}{
Provide an overview of currency availabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$availability_overview(assets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assets}}{the \link{Asset_Portfolio} providing all relevant asset information}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \linkS4class{tbl_df} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EIOPA_RFR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
