% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileWriter.R
\docType{class}
\name{FileWriter}
\alias{FileWriter}
\title{File Writer Object}
\format{
An \code{\link{R6Class}} generator object
}
\description{
AAM Credit ESG File Writer Object
}
\author{
Daniel Hohmann
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileWriter$new()}}
\item \href{#method-write}{\code{FileWriter$write()}}
\item \href{#method-write_final_output}{\code{FileWriter$write_final_output()}}
\item \href{#method-clone}{\code{FileWriter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new FileWriter instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWriter$new(type = "csv", csv_locale = "US", folder = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a string (csv or xlsx) specifying the file type}

\item{\code{csv_locale}}{a string (US or DE) specifying the csv file seperator and decimal convention}

\item{\code{folder}}{an optional string which can pre-define the output folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write a data.frame to a csv or xlsx file based on the pre-defined file type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWriter$write(data, filename, folder = NA, add_filename_extension = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \linkS4class{tbl_df} object to be written to the file}

\item{\code{filename}}{a character string specifying the filename of the file to be written}

\item{\code{folder}}{a character string specifying the folder of the file to be written (if NA no additional folder is added to the output path)}

\item{\code{add_filename_extension}}{boolean specifying if the filename extension should be added in case it is missing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_final_output"></a>}}
\if{latex}{\out{\hypertarget{method-write_final_output}{}}}
\subsection{Method \code{write_final_output()}}{
Write a data.frame to a csv of xlsx file based on the pre-defined file type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWriter$write_final_output(
  data_portfolio,
  data_asset,
  filename,
  folder = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_portfolio}}{a \linkS4class{tbl_df} object to be written in the first sheet with the information on portfolio level}

\item{\code{data_asset}}{a \linkS4class{tbl_df} object to be written in the second sheet with the information on asset level}

\item{\code{filename}}{a character string specifying the filename of the file to be written}

\item{\code{folder}}{a character string specifying the folder of the file to be written (if NA no additional folder is added to the output path)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWriter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
