% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLdb.R
\docType{class}
\name{SQLdb}
\alias{SQLdb}
\title{SQLite database Object}
\format{
An \code{\link{R6Class}} generator object
}
\description{
SQLite database Object
}
\author{
Daniel Hohmann
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SQLdb$new()}}
\item \href{#method-listTables}{\code{SQLdb$listTables()}}
\item \href{#method-disconnect}{\code{SQLdb$disconnect()}}
\item \href{#method-clone}{\code{SQLdb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Open the database connection. A new data will be created in case dbpath is not an existing one
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLdb$new(dbpath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbpath}}{A string defining the path and file name of the database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listTables"></a>}}
\if{latex}{\out{\hypertarget{method-listTables}{}}}
\subsection{Method \code{listTables()}}{
List all tables in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLdb$listTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string giving the table names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLdb$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLdb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
