% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\docType{class}
\name{YieldCurve}
\alias{YieldCurve}
\title{Yield curve Object}
\format{
An \code{\link{R6Class}} generator object
}
\description{
An R class defining a yield curve object with corresponding
functionalities
}
\author{
Daniel Hohmann
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{YieldCurve$new()}}
\item \href{#method-print}{\code{YieldCurve$print()}}
\item \href{#method-calibrate_price_function}{\code{YieldCurve$calibrate_price_function()}}
\item \href{#method-spot_rates}{\code{YieldCurve$spot_rates()}}
\item \href{#method-clone}{\code{YieldCurve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new yield curve object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YieldCurve$new(terms, spotrates = NULL, zcb = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms}}{An integer vector providing the input terms}

\item{\code{spotrates}}{A double vector providing the spot rates corresponding
to the input terms}

\item{\code{zcb}}{A double vector providing the zero coupon bond prices
corresponding to the input terms (either spotrates or zcb must be
provided)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the yield curve object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YieldCurve$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate_price_function"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate_price_function}{}}}
\subsection{Method \code{calibrate_price_function()}}{
Calibrate the Smith-Wilson zero coupon bond proxy
price function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YieldCurve$calibrate_price_function(
  alpha = 0.2,
  UFR = 0.042,
  UFR_correction = TRUE,
  alpha_search = FALSE,
  alpha_range = c(0.05, 0.8),
  UFR_convergence_point = 60,
  UFR_precision = 1e-04,
  alpha_precision = 1e-06
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A double specifying the convergence rate (does not need to
be provided in case alpha_search = TRUE)}

\item{\code{UFR}}{A double specifying the ultimate forward rate}

\item{\code{UFR_correction}}{A Boolean specifying if the given UFR is annually
compounded (TRUE) or continuously compounded (FALSE)}

\item{\code{alpha_search}}{A Boolean specifying if the optimal convergence
rate according to the EIOPA methodology is to be derived}

\item{\code{alpha_range}}{A double vector specifying the lower and upper
bound for the alpha search}

\item{\code{UFR_convergence_point}}{An integer specifying the term at which
the UFR is to be approximated}

\item{\code{UFR_precision}}{A double specifying the precision up to which the
UFR is to be approximated}

\item{\code{alpha_precision}}{A double specifying the precision up to which
alpha is to be optimized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spot_rates"></a>}}
\if{latex}{\out{\hypertarget{method-spot_rates}{}}}
\subsection{Method \code{spot_rates()}}{
Calculate spot rates for a given vector of terms given
the price function calibration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YieldCurve$spot_rates(terms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms}}{An integer vector specifying the terms}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A double vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YieldCurve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
